% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{coa-package}
\alias{coa}
\alias{coa-package}
\title{coa: Compound Assignment Operators}
\description{
The package will provide the operator \verb{\%+=\%}, which increments a variable by a given value:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- 1
x \%+=\% 10
x
#> [1] 11
}\if{html}{\out{</div>}}

This is called a "compound assignment" operator
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xFoxFelix/coa}
  \item Report bugs at \url{https://github.com/xFoxFelix/coa/issues}
}

}
\author{
\strong{Maintainer}: Felix Fox \email{meinebestimmtechteEmail@mail.de}

}
