% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%c=\%}
\alias{\%c=\%}
\title{Concatinate assignment}
\usage{
l \%c=\% r
}
\arguments{
\item{l}{variable to modify. atomic vector}

\item{r}{what to add. atomic vector}
}
\description{
Simple Concatinate assignment operator
}
\examples{
x <- 1:5
y <- c("a", "b", "c")
x \%c=\% 6:10
y \%c=\% c("d", "e")
}
