% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%-=\%}
\alias{\%-=\%}
\title{Compound Assignment Subtract Operator}
\usage{
t \%-=\% s
}
\arguments{
\item{t}{The left hand side that gets updated.}

\item{s}{The right hand side that gets subtracted from the left hand side.}
}
\description{
Subtract something from the left hand side in place.
}
\examples{
x <- 3
x \%-=\% 2
print(x)
}
