% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Addition Operator}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{LHS}

\item{s}{RHS}
}
\description{
This operator adds the RHS to the LHS and updates the LHS.
}
\examples{
x <- 1
x \%+=\% 10
x
}
