% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{minus-minus}
\alias{minus-minus}
\alias{--}
\title{Decrement by one}
\usage{
`--`(l)
}
\arguments{
\item{l}{variable to modify. class: numeric}
}
\description{
equivalent of l \%-=\% 1
}
\examples{
x <- 1:10
y <- 20
`--`(x)
`--`(y)
}
