% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%-=\%}
\alias{\%-=\%}
\title{Subtract assignment}
\usage{
l \%-=\% r
}
\arguments{
\item{l}{variable to modify. class: numeric}

\item{r}{by what value to modify. class: numeric}
}
\description{
Simple subtract assignment operator
}
\examples{
x <- 1:10
y <- 5
x \%-=\% 10
x \%-=\% y
y \%-=\% 10
}
