% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Compund Assignment Add Operator}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{The left hand side that gets updated}

\item{s}{The right hand side that gets updated}
}
\description{
Add something to the left hand side in place
}
\examples{
x <- 1
x \%+=\% 10
print(x)
}
